fix: block detail screen

+++ apps/processing/connectors/sources/web3/src/main/kotlin/com/ethvm/kafka/connect/sources/web3/sources/ParityBlocksSource.kt
@@ -33,7 +33,7 @@ class ParityBlocksSource(
 
   override val partitionKey: Map<String, Any> = mapOf("model" to "block")
 
-  private val blockTimestamps = sortedMapOf<BigInteger, Long>()
+  private val blockTimestamps = sortedMapOf<BigInteger, Int>()
   private val timestampsLock = AtomicBoolean(false)
 
   override fun fetchRange(range: LongRange): List<SourceRecord> {
@@ -52,7 +52,7 @@ class ParityBlocksSource(
 
             // record timestamp
 
-            val blockTimestamp = block.timestamp.toLong()
+            val blockTimestamp = block.timestamp.toInt()
             blockTimestamps[blockNumberBI] = blockTimestamp
 
             // uncles
@@ -103,7 +103,7 @@ class ParityBlocksSource(
 
       val blockTime = when (val prevTimestamp = blockTimestamps[blockNumber.minus(BigInteger.ONE)]) {
         null -> null
-        else -> block.timestamp.toLong() - prevTimestamp
+        else -> block.timestamp.toInt() - prevTimestamp
       }
 
       val blockRecord = block
